requirements.txt
azure-storage-blob==12.19.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
pyodbc==5.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Werkzeug==3.0.1
WTForms== 3.1.2
jinja2==3.1.3
itsdangerous==2.1.2
